refactor: Eliminate code duplicity, down to 0%
