--- src/screens/AddDebtorScreen/style.ts
@@ -1,10 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const addDebtorStyles = StyleSheet.create({
-  headerContainer: {
-    marginBottom: 20,
-    marginTop: 20,
-  },
-});
-
-export default addDebtorStyles;
