--- src/screens/AddTransactionsScreen/style.ts
@@ -1,31 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const addTransactionStyles = StyleSheet.create({
-  headerContainer: {
-    marginBottom: 20,
-    marginTop: 20,
-  },
-  addButtonContainer: {
-    marginBottom: 10,
-  },
-  submitButtonContainer: {
-    marginTop: 5,
-    marginBottom: 15,
-  },
-  dateButtonContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: 40,
-    width: 40,
-    borderRadius: 5,
-    borderWidth: 2,
-    marginRight: 10,
-  },
-  dateContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-});
-
-export default addTransactionStyles;
