+++ src/screens/ChooseCurrencyScreen/index.tsx
@@ -1,4 +1,4 @@
-import {ScrollView, TextInput, TouchableOpacity, View} from 'react-native';
+import {ScrollView, TextInput, View} from 'react-native';
 import React from 'react';
 import styles from './style';
 import Icon from '../../components/atoms/Icons';
@@ -7,6 +7,7 @@ import useChooseCurrency from './useChooseCurrency';
 import PrimaryView from '../../components/atoms/PrimaryView';
 import PrimaryText from '../../components/atoms/PrimaryText';
 import textInputStyles from '../../styles/textInput';
+import CurrencySymbolPicker from '../../components/molecules/CurrencySymbolPicker';
 
 const ChooseCurrencyScreen = () => {
   const {
@@ -65,35 +66,11 @@ const ChooseCurrencyScreen = () => {
 
       <ScrollView showsVerticalScrollIndicator={false}>
         <View style={styles.currencyMainContainer}>
-          {filteredCurrencies.map(currency => (
-            <TouchableOpacity
-              key={currency.code}
-              onPress={() => handleCurrencySelect(currency)}>
-              <View
-                style={[
-                  styles.currencyContainer,
-                  {
-                    backgroundColor:
-                      selectedCurrency === currency
-                        ? colors.accentGreen
-                        : colors.primaryText,
-                    borderColor: colors.secondaryText,
-                  },
-                ]}>
-                <View style={styles.symbolContainer}>
-                  <PrimaryText style={{color: colors.buttonText, fontSize: 20}}>
-                    {currency.symbolNative}
-                  </PrimaryText>
-                  <PrimaryText style={{color: colors.buttonText, fontSize: 13}}>
-                    {currency.code}
-                  </PrimaryText>
-                </View>
-                <PrimaryText style={{color: colors.buttonText, fontSize: 10}}>
-                  {currency.name}
-                </PrimaryText>
-              </View>
-            </TouchableOpacity>
-          ))}
+          <CurrencySymbolPicker
+            filteredCurrencies={filteredCurrencies}
+            selectedCurrency={selectedCurrency}
+            handleCurrencySelect={handleCurrencySelect}
+          />
         </View>
       </ScrollView>
