+++ src/screens/EverydayTransactionScreen/style.ts
@@ -1,10 +1,6 @@
 import {StyleSheet} from 'react-native';
 
 const everydayTransactionStyles = StyleSheet.create({
-  headerContainer: {
-    marginBottom: 20,
-    marginTop: 20,
-  },
   noTransactionContainer: {
     alignItems: 'center',
     justifyContent: 'center',
