+++ src/styles/main.ts
@@ -1,5 +1,10 @@
 import {StyleSheet} from 'react-native';
 
-const mainStyles = StyleSheet.create({});
+const mainStyles = StyleSheet.create({
+  headerContainer: {
+    marginBottom: 20,
+    marginTop: 20,
+  },
+});
 
 export default mainStyles;
